📦 Publish SDKs (#2930)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-vue@0.11.2

### Patch Changes

-   2821f68: Fix CSS imports:

- add a `/nuxt` subpath export for a Nuxt module. Currently, it handles
importing CSS, and is used like so:

    ```ts
    // nuxt.config.js
    export default defineNuxtConfig({
      modules: ['@builder.io/sdk-vue/nuxt'],
    });
    ```

- bring back `/css` subpath export. This is used internally by the Nuxt
module, but also allows more flexibility for customers to import CSS as
they see fit.
- remove CSS imports from server bundles, as they cause errors due to
invalid extensions.
    -   add `sideEffects` array to package.json for webpack.

## @builder.io/sdk-react-nextjs@0.11.2



## @builder.io/sdk-qwik@0.11.2



## @builder.io/sdk-react@0.11.2



## @builder.io/sdk-react-native@0.11.2



## @builder.io/sdk-solid@0.11.2



## @builder.io/sdk-svelte@0.11.2

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>