📦 Publish SDKs (#2969)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-qwik@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-react@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-react-native@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-solid@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-svelte@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

## @builder.io/sdk-vue@0.12.5

### Patch Changes

- e7f6db6: Fix: sigfault crash when using SDK in Node v20 + M1 Macs.
Skip usage of `isolated-vm` in those environments.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>