Update heavy-cycles-approve.md (#2976)
## Description

Add a short description of what changes you made, why you made them, and
any other context that you think might be helpful for someone to better
understand what is contained in this pull request. This sort of
information is useful for people reviewing the code, as well as anyone
from the future trying to understand why changes were made or why a bug
started happening.

_Screenshot_
If relevant, add a screenshot or two of the changes you made.