📦 Publish SDKs (#2986)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-qwik@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-react@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-react-native@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-solid@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-svelte@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

## @builder.io/sdk-vue@0.13.0

### Minor Changes

- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only
'v3' is now allowed.

### Patch Changes

-   2b67586: Fix: TypeScript types for all exports.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>