Fix: React sdk edge bundle (#3023)
## Description

- add `/index.ts` files for all components
- move all `/edge`, `/browser`. `/node` subpath exports to `/bundle/*`

React SDK:
- lazyload Content component in edge runtime using `React.lazy`
- add a `/edge` subpath export pointing to a new export file that's a
modified version of `blocks-exports.ts`. It re-exports everything as
normal, but wraps the `Content` export to toggle between edge/browser
exports at runtime. This guarantees that the edge Content component,
which is lazy loaded, does not make it to the browser.
- thanks to the above, remove need for manual workaround of edge +
browser bundles in React SDK when in edge envs.