📦 Publish SDKs (#3403)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.14.30

### Patch Changes

- 7ae4170: Feature: add `@builder.io/sdk-react-nextjs/node/init` entry
point with `initializeNodeRuntime` export that sets the IVM instance.

This import should be called in a server-only location, such as _inside_
of a Next.js async `Page` component.

    ```tsx
    export default async function Page(props: MyPageProps) {
      // NOTE: the import must be inside the Page component itself.
      const { initializeNodeRuntime } = await import(
        "@builder.io/sdk-react-nextjs/node/init"
      );
      initializeNodeRuntime();

      // rest of your logic...
    }
    ```

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>