fix: resolver approach for ssr in angular and `onInit` (#3437)
## Description

Uses `onInit` in both CSR and SSR and `onMount` & `onUpdate` got shifted
to only run on the browser. This uses the resolver approach for fetching
content during SSR. This also fixes `blocks-wrapper` not respecting
children inside and moving them out.

_Screenshot_
If relevant, add a screenshot or two of the changes you made.