📦 Publish SDKs (#3468)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@2.2.7

### Patch Changes

- b965695: Fix: reverts `v2.2.5` change to `userAttributes` parsing
logic, as it caused breaking changes in certain cases.

## @builder.io/react@4.0.2

### Patch Changes

-   Updated dependencies [b965695]
    -   @builder.io/sdk@2.2.7

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>