📦 Publish SDKs (#3482)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@2.2.9

### Patch Changes

-   b7c00cf: Silence errors from non essential query params parsing

## @builder.io/react@4.0.4

### Patch Changes

- b7c00cf: Fix SSR hydration issues with non-hover animated builder
blocks
-   Updated dependencies [b7c00cf]
    -   @builder.io/sdk@2.2.9

## @builder.io/sdk-angular@0.2.6

### Patch Changes

- 22a3865: Fix: Symbol infinite re-render and editor overlay not showing
up.
-   a6c453f: Fix: support Visual Editing for Interactive Elements

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>