📦 Publish SDKs (#3489)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@3.0.0

### Major Changes

- bc1d409: Fix: Reintroduced `JSON.stringify(userAttributes)` change to
standardize parsing logic and preserve strings. This is a breaking
change as it doesn't require manual stringification of `userAttributes`
values. Ensure that attributes are not manually stringified before
passing them to avoid potential issues.

### Patch Changes

-   1586519: Fix: remove `next: { revalidate: 1 }` in SDKs fetch

## @builder.io/react@5.0.0

### Major Changes

- bc1d409: Breaking Change 🧨: `userAttributes` now is parsed as an
object - `JSON.stringify(userAttributes)` which preserves strings. Users
no longer need to manually stringify anything unless they have
explicitly included it in custom targeting attributes.

    For example,

    ```js
    userAttributes: {
      stringWithStrs: ["a", "c"];
    }
    ```

    used to work as well as this,

    ```js
    userAttributes: {
      stringWithStrs: ["'a'", "'c'"];
    }
    ```

but now its not needed to manually stringify strings. This change was
needed to preserve data types and strings so previously when we passed,

    ```js
    userAttributes: {
      stringWithNums: ["1", "2"];
    }
    ```

they were actual string numbers but we failed to parse it because we
were not preserving the strings and users had to perform manual
stringification hacks like `"'1'"` to achieve correct result. With this
change stringified numbers/bools etc will work out of the box as
expected showing less room for randomness.

### Patch Changes

-   1586519: Fix: remove `next: { revalidate: 1 }` in SDKs fetch
-   Updated dependencies [bc1d409]
-   Updated dependencies [1586519]
    -   @builder.io/sdk@3.0.0

## @builder.io/sdk-react-native@2.0.4

### Patch Changes

-   5a6cd39: Fix: ignore animations (avoid crashes when they are used).

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>