📦 Publish SDKs (#3508)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@3.0.1

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/react@5.0.1

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions
-   Updated dependencies [69859d4]
    -   @builder.io/sdk@3.0.1

## @builder.io/sdk-angular@0.2.9

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/sdk-react-nextjs@0.16.6

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/sdk-qwik@0.16.6

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/sdk-react@2.0.9

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/sdk-react-native@2.0.9

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions
- f1527ee: Allow gap, rowGap, columnGap in react native strict mode
styles

## @builder.io/sdk-solid@2.0.9

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

## @builder.io/sdk-svelte@2.0.9

### Patch Changes

- 69859d4: serialize functions for registering plugins so you can have
showIf on fields as functions

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>