📦 Publish SDKs (#3514)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-native@2.0.11

### Patch Changes

-   077cae9: strict validation fixes for react native

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>