📦 Publish SDKs (#3570)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react@5.0.3

### Patch Changes

-   20953a8: Disable localization on dynamic container inputs

## @builder.io/sdk-angular@0.2.13

### Patch Changes

- a44d73b: Fix: add `types` `exports` key to fix TS types support for
projects in `bundler` mode.

## @builder.io/sdk-react-nextjs@0.16.10

### Patch Changes

- a44d73b: Fix: add `types` `exports` key to fix TS types support for
projects in `bundler` mode.

## @builder.io/sdk-qwik@0.16.10

### Patch Changes

- a44d73b: Fix: add `types` `exports` key to fix TS types support for
projects in `bundler` mode.

## @builder.io/sdk-react@2.0.14

### Patch Changes

- a44d73b: Fix: add `types` `exports` key to fix TS types support for
projects in `bundler` mode.

## @builder.io/sdk-svelte@2.0.14

### Patch Changes

- a44d73b: Fix: add `types` `exports` key to fix TS types support for
projects in `bundler` mode.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>