📦 Publish SDKs (#3646)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@3.0.3

### Patch Changes

- d403fca: Adds `apiEndpoint` prop to `builder.get()` and
`builder.getAll()` with options `'content'` and `'query'`. It dictates
which API endpoint is used for the content fetching.
    Defaults to `'query'`

## @builder.io/react@5.0.6

### Patch Changes

- d403fca: Adds `apiEndpoint` prop to `builder.get()` and
`builder.getAll()` with options `'content'` and `'query'`. It dictates
which API endpoint is used for the content fetching.
    Defaults to `'query'`
-   Updated dependencies [d403fca]
    -   @builder.io/sdk@3.0.3

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>