📦 Publish SDKs (#3682)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@3.0.6

### Patch Changes

- b5dd732: Feature: start sending accurate npm package version
information to the Visual Editor

## @builder.io/react@5.0.10

### Patch Changes

- b5dd732: Feature: start sending accurate npm package version
information to the Visual Editor
-   Updated dependencies [b5dd732]
    -   @builder.io/sdk@3.0.6

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>