📦 Publish SDKs (#3712)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react@2.0.27

### Patch Changes

- 800457c: Feature: add `/node/setIvm` export path. It provides a
`setIvm` function that can be used to provide the SDK with an instance
of the `isolated-vm` module. This is needed in Node.js environments for
data bindings to work when `/node/init` can't be used.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>