📦 Publish SDKs (#3715)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-qwik@0.16.21

### Patch Changes

-   a25a805: Fix: build to work in preview mode

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>