[smartling][deleteFeatureFlag] Removing feature flag check (#3728)
## Description
The feature flag was temporarily added while users migrate their locales
from smartling to builder. We have communicated with the users. The
migration is completed and it has been turned off since long time. It is
now safe to delete it.

Once the pr is merged and plugin is deployed, will delete the feature
flag from launch darkly.