📦 Publish SDKs (#3741)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@5.0.0

### Major Changes

- f12f43d: Breaking Change: Use `/query` instead of `/content` for API
calls. This change fixes a symbol rendering issue introduced in
<https://github.com/BuilderIO/builder/pull/3681>, which was included in
the 4.0.0 release.

## @builder.io/react@7.0.0

### Major Changes

- f12f43d: Breaking Change: Use `/query` instead of `/content` for API
calls. This change fixes a symbol rendering issue introduced in
<https://github.com/BuilderIO/builder/pull/3681>, which was included in
the 6.0.0 release.

### Patch Changes

-   Updated dependencies [f12f43d]
    -   @builder.io/sdk@5.0.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>