[sfccPlugin] Validation on sfcc config (#3767)
## Description

In the SFCC plugin, previously, settings were not validated after being
saved. This could lead to confusion for users, as any incorrect
configuration would only surface as errors when the plugin was used
later.

To improve user experience and reduce failure points, we've implemented
a "fail-fast" approach. Now, as soon as users enter the configuration
details, an authentication call is triggered to validate the settings
immediately. This ensures that any issues are identified and
communicated to the user upfront.

**Loom**
Before loom:
https://www.loom.com/share/8d4af13e5b184465a5a53369e083de34

After loom:
https://www.loom.com/share/35f74aa2d6414668bf192dcf694083be